Добавляет модуль
